Fixed grade_histrogram() so that it returns results for display in Staff Debug Info.
grade_histogram() was returning no results if any of the StudentModule rows had grade == Null.
However, grade is Null if the student has only viewed a problem and not attempted it. This
meant that histograms were almost never being displayed (since the probability that everyone
who has viewed the problem has also attempted it is effectively 0).

LMS-2209